Fix misra

+++ board/safety.h
@@ -84,7 +84,7 @@ int get_addr_check_index(CAN_FIFOMailBox_TypeDef *to_push, AddrCheckStruct addr_
   return index;
 }
 
-void update_counter(AddrCheckStruct addr_list[], int index, int counter) {
+void update_counter(AddrCheckStruct addr_list[], int index, uint8_t counter) {
   if (index != -1) {
     if (addr_list[index].check_counter) {
       uint8_t expected_counter = (addr_list[index].last_counter + 1U) % addr_list[index].max_counter;
